fix(useField): Made useField Hook Reactive
This is recommended to do by Vue and removes most `.value`s from e.g. the tests and removed a forceRef hack.

When we had to wrap the fields in an array that was meant to be reactive, thus had to use the reactive on refs while trying to push fields.

Co-Authored-By: Carol Soliman <17387510+carsoli@users.noreply.github.com>