fix(useField): fix hack to of using `isGroup` to avoid having `label` as wrapper
when `label` is used, focus is detected when use clicks on any part of the component.
when using KtSingleSelect for example (we don't control focus handling), focus is detected twice (once on the el-select element and once within the label. this leads to inability to open the dropdown when we click the input itself because the dropdown opens then closes (fousing twice makes it toggle back to closed state)
passing down an option isComponent, allows the user to define any type of wrapper without having to have a `fieldset` by using isGroup hack
credits to Florian Wendelborn for finding the source of the problem