wip:feature(useField): KtFieldDateRange
wip:
end-start placeholders depend on translation context
may decide to make the shortcuts validation logic. more complex if we decide to support string shortcuts for the range picker.
refact: moved common logic between date and picker to constants and utils file
renamed the props that are hard-coded to el-ui to be prefixed by EL_ instead of them being related to the component name
moved affix icon logic for dates to mounted hook since it doesn't need to be in a watcher callback
style changes and additions for date components