fix(#268): Make Borders Re-appear by Removing no-outline-or-border Mixin
The idea of the mixin is reasonable, but its effects are rather unpredictable as the production build gets shuffled around differently than the development build and thus doesn’t show that some borders might be missing.

I’ve replaced all occurrences of it with what worked best in the situation.