fix(KtModal): does not show if isOpen prop is set to true initially
unless it changes and retriggers the watcher, the isOpen value
was dismissed in the first iteration
which led to a bug in announcementModal usages, where `isOpen` was true
and never changed throughout the lifecycle of the component.

Fix was to make the watcher immediate, keep track of the initial value
of `isOpen` and wait for tippy to be created and then show it
if the initial value was true.

Co-Authored-By: Florent Gallois<flgallois@gmail.com>