fix(KtTable): use attrs instead of domProps for specific attributes
Vue offers two different parameters to pass html attributes into rendered
elements, `attrs` and `domProps`. But they are not interchangable, in fact each seems to accept
a predefined set of attributes and will do nothing if the user passes an
attribute that the paramenter does not accept. It is unclear what the
logic behind this is, so the changes here have been found by experimentation