hack(vue-use-tippy): make sure tippy is always a function
In certain situations `import tippy from "tippy.js"` will not produce
a function, because tippy will be imported as cjs when it should be
imported as esm.

This hack tries to ensure correctness at runtime since a proper fix
at build time has not been found and we prefer to keep both builds as
export options.

Co-Authored-By: Florian Wendelborn <1133858+FlorianWendelborn@users.noreply.github.com>